src/requirements.txt
PyGithub==1.55
semver==2.13.0